requirements.txt
polars==0.20.29
pip==24.0
wheel==0.42.0
setuptools==69.1.0
numpy==1.26.4
lightfm==1.17
scipy==1.13.1
pandas==2.2.2
scikit-learn==1.5.0
pytz==2024.1
threadpoolctl==3.5.0
joblib==1.4.2
python-dateutil==2.9.0.post0
certifi==2024.2.2
requests==2.32.2
urllib3==2.2.1
six==1.16.0
idna==3.7
loguru==0.7.2
dill==0.3.8
catboost==1.2.5
pillow==10.3.0
packaging==24.0
matplotlib==3.9.0
plotly==5.22.0
tenacity==8.3.0
graphviz==0.20.3
contourpy==1.2.1
fonttools==4.52.1
pyparsing==3.1.2
cycler==0.12.1
kiwisolver==1.4.5
pydantic==2.7.1
fastapi==0.111.0
faiss-cpu